% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the location of accidents in a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that represents the state number.}

\item{year}{An integer or string indicating the year.}
}
\value{
Plots a map of the state, where the fatalities are the dots on the map.
}
\description{
\code{fars_map_state} is a function that creates a map with the fatalities in a US State
during a year according to the Fatality Analysis Reporting System (FARS) database.
}
\details{
This function takes the state number and a year as inputs and calls the supporting
   functions \code{make_filename()} and \code{fars_read()}.
}
\note{
The \code{state.num} parameter of this function must be an integer or a string with
   number. This function stops execution with an error message if \code{state.num} is
   invalid. If there hasn't been any accident in that year, this function produces a message.
}
\examples{
\dontrun{
fars_map_state(30, 2013)
fars_map_state(56, 2013)
fars_map_state(70, 2014) ## Error: "invalid STATE number: 70"
}

}
