% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of FARS data for the years in the parameter \code{years}.
}
\description{
\code{fars_read_years} is a function that reads Fatality Analysis Reporting System (FARS)
data for multiple years.
}
\details{
This function processes each element of a vector or list of years. To generate a
   file name with the year this function calls the supporting function \code{make_filename()},
   and then calls the supporting function \code{fars_read()}, to read the FARS data file
   with  the generated file name. If the file exists, the function creates a new column
   (year), then selects the month and year columns, and adds them to an R list.
}
\note{
If the file doesn't exit, the function gives a warning message, and adds nothing
   (NULL) to the list.
}
\examples{
\dontrun{
fars_read_years(c(2014, 2015, 2016))
fars_read_years(c("2014", "2015", "2016"))
}

}
